# docs/contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: LexFrame Stage 8 API
  version: 0.8.0
  description: Canonical HTTP contract for the LexFrame stage 8 execution core, embedded builder, approval inbox, artifact handling, delivery and notifications.
servers:
  - url: http://localhost:3100
paths:
...
  /document-types:
    get:
      operationId: listDocumentTypes
      tags: [Document Authoring]
      responses:
        "200":
          description: Document types.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentTypeSummary"
    post:
      operationId: createDocumentType
      tags: [Document Authoring]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentTypeRequest"
      responses:
        "200":
          description: Created document type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentTypeDetail"
  /document-types/{id}:
    get:
      operationId: getDocumentType
      tags: [Document Authoring]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document type detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentTypeDetail"
    patch:
      operationId: updateDocumentType
      tags: [Document Authoring]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentTypeRequest"
      responses:
        "200":
          description: Updated document type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentTypeDetail"
  /document-structures:
    get:
      operationId: listDocumentStructures
      tags: [Document Authoring]
      responses:
        "200":
          description: Document structures.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentStructureRecord"
    post:
      operationId: createDocumentStructure
      tags: [Document Authoring]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentStructureRequest"
      responses:
        "200":
          description: Created document structure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentStructureRecord"
  /document-structures/{id}:
    patch:
      operationId: updateDocumentStructure
      tags: [Document Authoring]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentStructureRequest"
      responses:
        "200":
          description: Updated document structure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentStructureRecord"
...
  /document-validations/{id}:
    get:
      operationId: getDocumentValidationReport
      tags: [Document Validation]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Validation report detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentValidationReportDetail"
  /document-validations/{id}/recheck:
    post:
      operationId: recheckDocumentValidation
      tags: [Document Validation]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecheckDocumentValidationRequest"
      responses:
        "200":
          description: Recomputed validation report.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentValidationReportDetail"
...
  /workflow-runtime/document-validation/execute:
    post:
      operationId: executeWorkflowRuntimeDocumentValidation
      tags: [Workflow Runtime]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowRuntimeDocumentValidationExecuteRequest"
      responses:
        "200":
          description: Runtime validation execution result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentValidationReportDetail"
...
components:
...
  schemas:
...
    Stage7Object:
      type: object
      additionalProperties: true
...
    DocumentTypeSummary:
      allOf:
        - $ref: "#/components/schemas/Stage7Object"
    DocumentTypeDetail:
      allOf:
        - $ref: "#/components/schemas/Stage7Object"
    CreateDocumentTypeRequest:
      allOf:
        - $ref: "#/components/schemas/Stage7Object"
    UpdateDocumentTypeRequest:
      allOf:
        - $ref: "#/components/schemas/Stage7Object"
    DocumentStructureRecord:
      allOf:
        - $ref: "#/components/schemas/Stage7Object"
    CreateDocumentStructureRequest:
      allOf:
        - $ref: "#/components/schemas/Stage7Object"
    UpdateDocumentStructureRequest:
      allOf:
        - $ref: "#/components/schemas/Stage7Object"
...
    DocumentValidationReportDetail:
      allOf:
        - $ref: "#/components/schemas/Stage7Object"
    RecheckDocumentValidationRequest:
      allOf:
        - $ref: "#/components/schemas/Stage7Object"
...
    WorkflowRuntimeDocumentValidationExecuteRequest:
      allOf:
        - $ref: "#/components/schemas/Stage7Object"
...